Rename frontend/src/api/outputs/ to api/visualizations/ and update imports
