+++ frontend/src/api/visualizations/Outputs.ts
